fix: fix booking page login bug 페이지마다 로그인 안되는 버그 수정
